Setting more stuff up. Player auto aim. Player shoot. Enemy move to player. Enemy auto shoot. And much more
Lots more setting up stuff.

Player auto aims to closest enemy.
Player shoots when they have a closest enemy on a loop for the reload time.

Enemies move to the player without any need for perception.
Enemies auto shoot when within a certain range of the player and will not shoot when they are too close to the player.

Hit detection for projectiles is just overlap for the static mesh for now.

Loads more stuff that I'm not writing in here because I'm too tired

Also added some content packages for special effects.